requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-tavily
chromadb
tavily-python
pypdf